+++ lib/import_path.dart
@@ -1,0 +1,14 @@
+/// Import Path search library.
+library import_path;
+
+export 'src/import_path_base.dart'
+    show
+        ImportPathStyle,
+        parseImportPathStyle,
+        ImportPathStyleExtension,
+        ImportToFind,
+        ImportToFindURI,
+        ImportToFindRegExp,
+        ImportPath;
+export 'src/import_path_parser.dart' show ImportParser;
+export 'src/import_path_scanner.dart' show ImportPathScanner;